[HOTFIX]: Manager now logs in to listProduct
