+++ src/main/resources/templates/listProduct.html
@@ -10,7 +10,7 @@
 <div class="text-center">
     <h1 class="text-6xl mb-6 font-extrabold">Welcome to HeyMart 👋</h1>
     <h2 class="text-3xl" th:text="'Hello, ' + ${username} + '! Your role is: ' + ${role}">Hello, username! role</h2>
-    <h1 class="text-3xl mb-6 font-extrabold">You're the user</h1>
+    <h1 class="text-3xl mb-6 font-extrabold">You're the manager</h1>
     <form action="/logout" method="post">
         <button type="submit" class="block w-full bg-blue-500 hover:bg-blue-700 text-white font-semibold py-2 px-4 rounded focus:outline-none focus:shadow-outline">Logout</button>
     </form>
